[ozone/linux] Default to built-in window decoration drawing
In Ozone/Chrome, DesktopWindowTreeHostMus installs a basic
NonClientFrameView instance (ClientSideNonClientFrameView), and relies
on Ash to fill in its content. This is not the case of Ozone/Linux
build.

For Chrome window decoration we are fine, since BrowserFrame
reimplements ::CreateNonClientFrameView [1], but for other Widgets
(including TaskManager), we end up without a proper window decoration.

Patch defaults DesktopWindowTreeHostMus::auto_update_client_area_ to 'false'
in Ozone/Linux builds, so that built-in default window decorations are provided.

We still need to make the Windows-looking theme of the decoration area
(light blue) to look better but this will be left over as a follow up.

Issue #260

[1] //chrome/browser/ui/views/frame/browser_frame.cc