Reset ui::ws::EventDispatcher's states when capture is lost.
This CL fixes a case, when a new native window is created and
all the events (including a button up event) go to that new window.

Before this CL, when a new window was created, the previous
window's EventDispatcher state was left with a button down
state, which caused events to be sent to a wrong target.

As of now, when a native window looses a capture, an EventDispatcher
is explicitly reset to make it possible to continue consuming events.

Issue #318

fixup! Reset ui::ws::EventDispatcher's states when capture is lost.

EventDispatcher -> EventProcessor.