Implement filter to run browser tests with Linux/mus build.
This filter disables all the crashing and failing tests and
leaves only the working ones. Some of the tests are broken
both with stock Chromium build and Chromium/mus build.

fixup! Implement filter to run browser tests with Linux/mus build.

It added failed tests on Mar/05 rebase to filter to skip them

TBR=msisov