[gpu-split][ozone/wayland] Use nested compositor approach when mus!=viz.
This is a first set of patches for gpu-split work.

It was decided to first use nested compositor approach to allow out
of browser process gpu service to work with wayland.

That is, there was no mean of communication between gpu wayland EGL and
browser side wayland connection.

The nested compositor approach resolves the problem by acting as
intermediate compositor, which creates wl_surface for a gpu client,
which is capable to attach egl buffer to that surface. Once eglSwapBuffers
happens, surface_commit method, which is implemented on nested compositor
side, does the trick to handle buffers data further to real host compositor
using real wl_surface representing a window on browser side.