[TEMP] Use discardable memory with Mash.
This is a workaround to resize crash in Mash: that is,
when Mash is used, we cannot properly resize the browser
window due a crash, which happens when browser tries to
access discardable memory. It was introduced to Mash
only when 8ff9c300b5f0a was merged.

fixup! [TEMP] Use discardable memory with Mash.

Fixed typo