[DoNotCarryForward] [ozone] Prepare Ozone desktop integration for Linux and other builds.
This cl adds an initial support for Ozone (Linux) builds by
utilizing DekstopWindowTreeHostPlatform design.

Actually, it looks very similar to DesktopWindowTreeHostX11 with
a difference that this design in this CL uses Ozone abstraction
layer to communicate with Wayland and other backends.

In overall, BrowserDesktopWindowTreeHostPlatform, which is actually
DesktopWindowTreeHostPlatform, is created and initialized, which triggers
creation of a PlatformWindow, which can be either Wayland, or X11, or other
ozone platform.

Also, DesktopScreenOzone class is added. This object inherits ScreenBase and
is meant to fetch displays data from underlaying ozone backends and provide
aura with necessary data.

Bug: 578890

Change-Id: I2d203cd6d1a6ddea0afe062059c6c1f9eecdafd3

This is a cherry-pick CL from
https://chromium-review.googlesource.com/c/chromium/src/+/1039567

Once uploaded, drop it in favor of merged cl.