fixup! Make window interactions (move, resize, double click and etc) working
Make sure that compound event filter is destroyed properly.

It may happen that OnClosed is closed, which leads to
DesktopNativeWidgetAura destructions. There is also another path
when DesktopNativeWidgetAura can be destroyed - CloseNow. That is,
if OnClosed has not been closed, CloseNow proceeds and destructs
DesktopNativeWidgetAura.

This patch makes it sure that the non_client_event_filter_ is destroyed
in any case.