[ozone/wayland/drm] Refactor and place gbm_device into shared place.
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and mades it functional when the gpu service is running
off the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, gbm_device.cc/h is moved into a common folder in the
ui/ozone/common/linux and renamed to gbm_device_linux.cc/h. What is more,
it is refactored in such a way that it can be used by both Wayland with
dmabuf approach and ozone/drm.

This is only compiled for wayland and drm backends, and does not bring
any additional load on other backends

Bug: 820047
Change-Id: I9eb4da474eb24d6d32fe355ff37b6d0806672fde