[ozone/wayland/drm] Place OverlayPlane and ScanoutBuffer into shared place.
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and mades it functional when the gpu service is running
off the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, overlay_plane.cc/h and scanout_buffer.h are
moved into a common folder in the ui/ozone/common/linux.
These are only compiled for wayland (downstream now) and drm backends,
and does not bring any aditional load on other backends

Bug: 820047
Change-Id: I1268eeb38e710e9a1c335def08fa94edf103095f