[ozone/wayland/drm] Refactor and place gbm_buffer into a shared place.
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and makes it functional when the gpu service is running
out of the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland for a dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, gbm_buffer.cc/h is moved into a common folder in the
ui/ozone/common/linux and renamed to gbm_buffer_linux.cc/h. What is more,
it is refactored in such a way that it can be used by both Wayland with
dmabuf approach and ozone/drm.

This is only compiled for wayland(downstream now) and drm backends, and
does not bring any additional load on other backends.

Bug: 820047
Change-Id: I10cf2bcb990aa6de8acb67e941cc4e774747682a