[ozone/common/drm] Avoid forward declaration on not declared deps.
Make sure forward declaration is not used to just avoid
declaring dependencies in the common/BUILD.gn.

That is, DrmDevice is a drm platform's class, which shouldn't be known
by the ozone/common.

To fix this, return GbmDeviceLinux instead and add AsGbmDeviceLinux method
to the DrmDevice class to return downcasted pointer to self, and move
ScanoutBufferGenerator to drm/gpu as long as it's not needed by the
Wayland with dmabuf approach at this stage.

Bug: 820047
Change-Id: I25703db41234275e0c0cb5aaefe0e61964a50e83