[ozone/wayland/IME] Handle backspace button.
This patch adds implenetation details to the delete surrounding
text and set surrounding text methods, which are used to handle
backspace buttons.

PS using vkb and physical keyboard at the same time are prone to bugs.
That is, when a text is typed both on the vkb and physical keyboard,
it can suddenly disappear.