[ozone/wayland] Ensure all the mojo calls happen on gpu main thread.
This cl fixes threading model: the proxy object is created on
a gpu main thread, whereas mojo ptr for a WaylandConnection is
set on a IO child thread. ScheduleSwap is also called on GpuMainThread,
but other methods can be called from IO threads.

To make a consitency, always use GpuMainThread's task runner.
This is also needed for a frame callback, which will be passed during
ScheduleSwap calls in the following cl.

Bug: 859012, 820047