[ozone/wayland] Add remove output functionality
This CL adds a new functionality to the existing WaylandOutputManager
and WaylandScreen classes. Now, it is possible to remove and
change primary displays on hot plugs and removes.

To test that functionality, a new test has also been added.

The logic that assigns output ids has also been changed.
Now, we are using numeric names that a Wayland compositor
announces during Global and GlobalRemove calls. That is,
each global object is assigned an unique numeric name, which can
be freely used as an id.

Bug: 890276, 875161
Change-Id: I775bdd40a946c9f0c5dde89eba6e36c4615e1f21
Reviewed-on: https://chromium-review.googlesource.com/c/1254146
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#596155}