[ozone] Implement Clipboard for Ozone platforms.
This is a first patch, which implements basic text
support for the Ozone platforms.

The flow is asynchronous and based on Requests. That is,
whenever a ClipboardOzone receives a Read/Write/GetMime
call, it forwards it to the AsyncClipboardOzone, which
then creates a request (used for internal usage and holding
data filled by the ClipboardDelegate), calls to the delegate
and start an abort timer to make sure the request is not stalled.

What is more the selection data is cached whenever it's written
to a system clipboard. And whenever the chromium browser is the
selection owner, the cached data is used.

Change-Id: Idb2033edf357423e86b12740365e0c5fe69e303a