Implement filter to run browser tests with Linux/Ozone
It's verified with '--ozone-platform=headless'
This filter disables all the crashing and failing tests and
leaves only the working ones.

TBR=msisov

fixup! Implement filter to run browser tests with Linux/Ozone

It updated the filter for browser_tests with failed test cases
on build-bot.

fixup! Implement filter to run browser tests with Linux/Ozone

Exclude failing tests. The reason of failures is unknown.

All the tests must be removed from this filter and added again
after a check.